Remove deprecated classes since they're not referenced anywhere
User class, while not directly interfacing with the DB like it used to,
is still used for convenience for getting the current user without
having to create a bunch of variables (to store the object, and the
instances, etc.) to do so.

There are no hardcoded TenantObject(Instance)Methods anymore! Every
method is an instance of the Method object. So that's pretty cool.