Give us a way for disabling Login Page for development purposes
If you're really paranoid, I recommend removing this block of code
entirely in case the location where your configuration file is stored is
ever compromised.

The reason I added this is to debug MOCHA built-in method "Get Login
Token for Current User", during the transition from Properties to
Attributes and Relationships. Something broke, and now I have to fix
it...