If $value is a Date object, translate it to a value we can store in DB
The big question is, should we eventually rewrite the DB schema to store
Text Attributes as LONGTEXT (as are all attributes right now), but have
Boolean Attributes as BIT, Date Attributes as DATETIME, and Numeric
Attributes as INT?

This would necessitate several different AttributeValues tables, each
for the different type of data. Is this The Right Thing(tm) to do?