Provide a way to specify options on Report Columns
This also includes a very useful option right out-of-the-box, for
Relationship Report Field "Display as Count" will display the resulting
data as a clickable drill-down labeled with the number of instances
hiding behind it. For cases where the list of instances would be large
and unwieldy, a number with the option of drilling down is quite useful.

Not even Workday (as far as I know) implements this, at least not in
this way... I believe this feature is only available in Matrix reports?