Implement AttributeEditorPageComponent
We still need to Objectify the implementations of Page Components; for
example, Page Component.has Build Element Method ... which would
generate the markup graph for the element associated with the Page
Component. We would still need some client-side code that would take the
resulting graph and build HTML output though (which is what this does
now).