Do not render the standard footer (used for Tasks) on Reports
I know WD uses the standard footer for the "Print" button on Reports
that have an associated Report Print Layout (currently not implemented),
but I think that would look better as a Page Option instead. Again,
currently not implemented, but it would basically take the form of a
miniature toolbar floating on the right side of the page, vertically
aligned with the Instance (Task or Report) title.