Ensure that Objects and Instances get unique IDs per tenant and object
Well, that took long enough. Finally was able to re-write most of the
codebase to support object-unique Instance IDs (e.g. Obj$Inst, so object
5 can have inst 1, and so can object 12 - 5$1, 12$1...)

Object-unique Instance IDs make it easier for us to see how many
instances of a particular object exist in the tenant, and provide a
concrete example that 21$53 on our tenant is not the same as 21$53 on
their tenant.

We had to implement our own "auto increment" system for this, and I
think it might be a bit slow. Improvements and suggestions are greatly
appreciated.