Move PagePage functionality (Page Builder) into ExecuteInstancePage
Eventually, each Class will have its own `.has execute Page` which will
be responsible for building the Page to be displayed and executed. Until
then, we do it manually, but the only hard-coded thing that should be
left in the end should be Page.

Furthermore, each Page Component should have its own Build Element
Method that would be responsible for building the component(s) that are
eventually rendered on the Page. None of that should be hardcoded.