Execute method calls for Tasks that have them
We have to figure out when we should do this and when we shouldn't; UI
Tasks do not usually execute methods after they request parameters, but
other Tasks do.

For example, `Create Class` should ask for parameters, execute a Method
to create the instance, and then redirect to the `Edit Class` task for
the newly-created instance.

On the other hand, `Edit Class` should ask for parameters (which are
filled in by the redirection from `Create Class` in that case), but its
methods for updating the Class (Process Updates Method?) should only be
executed after the task "body" is displayed.

Perhaps UI Tasks that do not have Page Components should always execute
methods the first way, and those that DO have Page Components should
always execute methods the second way?